+++ Clases/Pesticida.cs
@@ -11,7 +11,7 @@ namespace Clases
         private int _idPesticida;
         private string _nombre;
         private string _tipo;
-        private short _pH;       
+        private double _pH;       
         private string _impacto;
 
         public int IdPesticida { get => _idPesticida; set => _idPesticida = value; }
@@ -19,11 +19,11 @@ namespace Clases
         public string Tipo { get => _tipo; set => _tipo = value; }
      
         public string Impacto { get => _impacto; set => _impacto = value; }
-        public short PH { get => _pH; set => _pH = value; }
+        public double PH { get => _pH; set => _pH = value; }
 
         public Pesticida() { }
 
-        public Pesticida(int idPesticida, string nombre, string tipo, short ph, string impacto)
+        public Pesticida(int idPesticida, string nombre, string tipo, double ph, string impacto)
         {
             IdPesticida = idPesticida;
             Nombre = nombre;
